阶段 D：语义分析扩展

+++ examples/basic/hello_world.c
@@ -7,9 +7,9 @@
 // Global Variables
 
 // Forward Declarations
-int main();
+long long main();
 
-int main() {
+long long main() {
   cn_rt_init();
 
   entry:
